+++ src/main/java/com/ufape/agiota/comunication/dto/Avaliacao/AvaliacaoResponse.java
@@ -1,12 +1,9 @@
 package com.ufape.agiota.comunication.dto.Avaliacao;
 
-import com.ufape.agiota.comunication.dto.agiota.AgiotaResponse;
-import com.ufape.agiota.comunication.dto.customer.CustomerResponse;
+
 import com.ufape.agiota.config.SpringApplicationContext;
 import com.ufape.agiota.negocio.enums.Avaliado;
-import com.ufape.agiota.negocio.enums.Status;
 import com.ufape.agiota.negocio.models.Avaliacao;
-import com.ufape.agiota.negocio.models.Borrowing;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import lombok.AllArgsConstructor;
@@ -15,13 +12,10 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.modelmapper.ModelMapper;
 
-import java.math.BigDecimal;
-import java.util.Date;
-
 @Getter @Setter @NoArgsConstructor @AllArgsConstructor
 public class AvaliacaoResponse {
 
-    private int nota;
+    private double nota;
     @Enumerated(EnumType.STRING)
     private Avaliado avaliado;
